Introduce KeplerTable class (#1112)
